Base code for twitter photos powered by fabric.io api client.

+++ app/src/main/java/com/binaryfork/onmap/network/Media.java
@@ -17,9 +17,9 @@ public interface Media {
 
     String getSiteUrl();
 
-    float getLatitude();
+    double getLatitude();
 
-    float getLongitude();
+    double getLongitude();
 
     boolean isVideo();
